% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_performance.R
\name{evaluate_performance}
\alias{evaluate_performance}
\title{Evaluate performance of a model in \code{MLWIC2} on your dataset}
\usage{
evaluate_performance(model_output, top5 = TRUE, label_tbl = speciesID)
}
\arguments{
\item{model_output}{The output of running \code{classify} after you have run 
\code{make_output} on this output.}

\item{top5}{Boolean variable for if you want to evaluate top-5 results for the model.
This will only work if you saved at least 5 guesses from the model in \code{classify} by 
specifying `top_n=5` or greater.}

\item{label_tbl}{A table with words to identify the classes in your dataset. If you are 
using the built in model, this is the speciesID data file that was downloaded with the package.
If you are writing your own `label_tbl`, you need to have columns called "class_ID" and "group_name".
For example, see head(speciesID)}
}
\description{
\code{evaluate_performance} calculates some basic summary statistics of model 
performance using \code{classify}. After running \code{classify}, you will need
to run \code{make_output} and have this output as your data frame to put into
this function.
}
