% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_output.R
\name{make_output}
\alias{make_output}
\title{Present output from a dataset classified by \code{MLWIC}}
\usage{
make_output(output_location, model_dir, output_name = "output.csv",
  saved_predictions = "model_predictions.txt", return_df = FALSE,
  top_n = "5", shiny = FALSE)
}
\arguments{
\item{output_location}{Absolute path where you want the output csv stored. This path
must exist on your computer.}

\item{model_dir}{Absolute path to the location where you stored the L1 folder
that you downloaded from github.}

\item{output_name}{Desired name of the output file. It must end in `.csv`}

\item{saved_predictions}{This is the file name where you stored predictions when you ran
\code{classify}. If you used the default in that function, you can use the default here.}

\item{return_df}{If TRUE, this will return the model output. This allows you to create a variable that 
is the model output.}

\item{top_n}{The number of guesses that you wanted classify to save. This needs to mach what you specified 
for top_n in \code{classify}}
}
\description{
\code{make_output} will make a clean csv presenting the results from your run
of \code{classify}. If you are planning to avoid typing absolute paths, you need to first
set your working directory to the location where you stored your images folder and data_info csv. 
If you set this function to run and assign it to a variable, you can specify \code{return_df=TRUE}
and produce a variable in your R session that is the model output.
}
