% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runShiny.R
\name{runShiny}
\alias{runShiny}
\title{A wrapper function to run Shiny Apps from \code{MLWIC2}.}
\usage{
runShiny(app = "classify")
}
\arguments{
\item{app}{The name of the app you want to run. The options are currently 
`classify` and `train`.}
}
\description{
A wrapper function to run Shiny Apps from \code{MLWIC2}.
}
