% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup}
\alias{setup}
\title{Setup your computer to run \code{MLWIC2}}
\usage{
setup(
  python_loc = "/anaconda3/bin/",
  conda_loc = "auto",
  r_reticulate = FALSE,
  gpu = FALSE,
  envname = "r-reticulate"
)
}
\arguments{
\item{python_loc}{The location of python on your machine. If you are
using a Macintosh, the default is the likely location.}

\item{conda_loc}{The location of conda. It is usually in the same folder as python}

\item{r_reticulate}{Logical. Do you have an environment called "r-reticulate" for which you have
installed Python packages previously and want to retain these packages. Default is FALSE.}

\item{gpu}{Logical. Do you want to use a GPU for classifying and training. (You must have
a GPU on your machine for this to work).}

\item{envname}{The name of the conda environment you'd like to set up. 
If you don't manage multiple environments, leave this as default}
}
\description{
\code{setup} installs necessary Python packages on your computer. You will
need to run this before running \code{classify} and \code{train}. If this is your first time using
this function, you should see additional documentation at https://github.com/mikeyEcology/MLWIC2.
If you follow the link to install Anacoda and you are using a Mac, python should be in the default location.
}
