% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_and_classify.R
\name{setup_and_classify}
\alias{setup_and_classify}
\title{Automatically set up your computer to run \code{MLWIC2} and run \code{classify}}
\usage{
setup_and_classify(
  python_loc = "/anaconda3/bin/",
  conda_loc = "auto",
  already_downloaded_model = FALSE,
  tensorflow_installed = FALSE,
  MLWIC2_already_setup = FALSE,
  model_dir = getwd(),
  os = c("Mac", "Windows", "Ubuntu"),
  model_type = c("species", "empty_animal"),
  input_file = NULL,
  directory = getwd(),
  path_prefix = getwd(),
  image_file_suffixes = c(".jpg", ".JPG"),
  recursive = TRUE,
  images_classified = FALSE,
  output_location = getwd(),
  output_name = "MLWIC2_output.csv",
  shiny = FALSE,
  print_cmd = FALSE
)
}
\arguments{
\item{python_loc}{The location of python on your machine. If you are
using a Macintosh, the default is the likely location.}

\item{conda_loc}{The location of conda. It is usually in the same folder as python}

\item{already_downloaded_model}{logical. If TRUE, you have already downloaded the model and 
will specify its location as `model_dir`.}

\item{tensorflow_installed}{logical. If TRUE, you have already downloaded tensorflow on 
your machine}

\item{MLWIC2_already_setup}{logical. If TRUE, you have already setup MLWIC.}

\item{model_dir}{Absolute path to the location where you stored the trained folder
that you downloaded from github. If you specified `already_downloaded_model=TRUE`, then this
is the location where you stored the trained model folder.}

\item{os}{The operating system on your computer. Options are "Mac", "Windows", "Ubuntu".}

\item{model_type}{Do you want the model to ID species (`species`) or just determine if images are empty or
containing animals (`empty_animal`)?}

\item{input_file}{The name of your input csv. It must contain a column called "filename"
and unless you are using the built in model, a column called "class" (which would be your species or group of species).
If you don't know what is in your images, it is easiest to not have in `input_file` and to specify
`images_classified == FALSE`.}

\item{directory}{The directory of your `input_file`.}

\item{path_prefix}{Path to where your images are stored. You need to specify this if 
you want MLWIC2 to `find_file_names`.}

\item{image_file_suffixes}{The suffix for your image files. Only specify this if you are 
using the `find_file_names` option. The default is .jpg files. This is case-sensitive.}

\item{images_classified}{logical. If TRUE, you have classifications to go along with these images
(and you want to test how the model performs on these images).}

\item{output_name}{Desired name of the output file. It must end in `.csv`}
}
\description{
\code{setup_and_classify} is designed to be a simpler option for setting up and using \code{MLWIC}
to classify images for those uers who are less comfortable in computing. It requires fewer steps,
but also loses some of the flexibility associated with the functions it replaces (\code{setup}, 
\code{make_input}, \code{classify}, and \code{make_output}). This function will work with either
the species model or the empty/animal model.
}
