% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speciesID-data.R
\docType{data}
\name{speciesID}
\alias{speciesID}
\title{Species/group ID for \code{MLWIC2}}
\format{a dataframe}
\usage{
data(speciesID)
}
\description{
This data is the `class_ID` identifier for each species or group of species
in the \code{MLWIC2} trained model.
}
\examples{
data(specisID)

}
\keyword{datasets}
