% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windows_helper.R
\name{windows_helper}
\alias{windows_helper}
\title{A helper function to make shiny work on Windoes computers}
\usage{
windows_helper(FUN, input)
}
\arguments{
\item{FUN}{# function to run}

\item{input}{input from shiny}
}
\description{
\code{windows_helper} is an internal function
}
