% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_metadata.R
\name{write_metadata}
\alias{write_metadata}
\title{Write model predictions from \code{MLWIC2} to image files}
\usage{
write_metadata(
  output_location = getwd(),
  output_name = "MLWIC2_output.csv",
  model_type = c("species", "empty_animal"),
  show_sys_output = FALSE
)
}
\arguments{
\item{output_location}{The path to the output file that you created with \code{classify} or \code{make_output}}

\item{output_name}{The name of the output file that you created with \code{classify} or \code{make_output}}

\item{model_type}{Did you run the model to ID species (`species`) or just determine if images are empty or
containing animals (`empty_animal`)?}

\item{show_sys_output}{logical. If TRUE, shows the output from the system command}
}
\description{
Uses `exiftool` to add metadata to your image files based on output from running
\code{classify} on your images. You must have ExifTool installed on your machine
for this function to work (https://exiftool.org/install.html). This function is designed
for those users who want to be able to view their classified images along with the classifications
from \code{MLWIC} in software like digiKam, MediaPro, and Lighthouse.
}
