% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_input.R
\name{make_input}
\alias{make_input}
\title{Create an input file to run \code{classify} or \code{train} in \code{MLWIC}}
\usage{
make_input(input_file = NULL, usingBuiltIn = TRUE,
  images_classified = FALSE, trainTest = FALSE, file_prefix = "",
  propTrain = 0.9, directory = getwd())
}
\arguments{
\item{input_file}{The name of your input csv. It must contain a column called "filename"
and unless you are using the built in model, a column called "class" (which would be your species or group of species).}

\item{usingBuiltIn}{logical. If TRUE, you are setting up a data file to classify images using
the built in model.}

\item{images_classified}{logical. If TRUE, you have classifications to go along with these images
(and you want to test how the model performs on these images).}

\item{trainTest}{logical. Do you want to create separate csvs for training and testing}

\item{file_prefix}{What you want to appear as the filename before the suffix. If you are
only creating a file to test the model, you could specify "test_" and your output file name
would be "test_image_labels.csv". If you specify `trainTest = TRUE`, your suffixes will automatically be
"_train.csv" and "_test.csv"}

\item{propTrain}{proportion of images you want for training. `1-propTrain` is the proportion
that will be used for testing the model.}

\item{directory}{Directory of your input csv. The default option is your working directory.
The file created by this function will be stored in this same directory.}
}
\description{
\code{make_inpu} will make a csv with the specifications necessary to either classify 
images or to train a new model. Your `input_file` must contain a column called "filename"
and a column called "species".
}
