% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_spaces.R
\name{remove_spaces}
\alias{remove_spaces}
\title{Remove spaces from all files within a directory}
\usage{
remove_spaces(path = getwd(), pattern = c(".JPG", ".jpg"))
}
\arguments{
\item{path}{The path to the image files whose names need to be changed. The default
is to use your current working directory.}

\item{pattern}{A vector containing the file name suffixes for which you want to
change the name.}
}
\description{
\code{remove_spaces} removes paces from file names, which is required to 
run MLWIC2 functions on these images. In the future, it is good practice to 
avoid putting spaces in any folder or file name.
}
